Move entrypoint to main.py, remove prefix overrides
